Create README.md
This Project is based on react, this is an website where the teachers can post the books, journals and assignment details. in backend i have been used the mongoDB for storing the data and mySQL for role based login page. Even i add a chatting space where teacher can message and student can reply.